pyproject.toml
[build-system]
requires = [
  "setuptools>=49",
]
build-backend = "setuptools.build_meta"

[project]
name = "pymmunomics"
version = "1.0.0"
description = "Package containing adaptive immune receptor analysis methods"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
authors = [
  {name = "Jasper Braun", email = "jasperbraun90@gmail.com"},
]
keywords = ["AIRR", "immunology", "immune repertoire"]
dynamic = ["dependencies"]

[project.urls]
repository = "https://github.com/JasperBraun/pymmunomics.git"

[tool.setuptools.dynamic]
dependencies = {file = ["requirements.txt"]}

[project.optional-dependencies]
dev = [
  "numpydoc>=1.5.0",
  "pytest>=7.3.1",
  "pytest-cov>=4.1.0",
  "sphinx>=7.0.1",
]
